doctor/requirements.txt
# Requirements for Render deployment
Flask==3.0.3
Flask-CORS==4.0.1
gunicorn==22.0.0
pymongo==4.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.3.0
setuptools==69.0.3
bcrypt==4.1.2

# Voice Dictation Dependencies
Flask-SocketIO==5.3.6
elevenlabs==0.2.26
websockets==11.0.3
python-socketio==5.9.0

# RAG (Retrieval-Augmented Generation) Dependencies
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0

# Background task scheduling
APScheduler==3.10.4